+++ match/src/main/java/cupid/recommend/query/param/RecommendQueryParam.java
@@ -29,8 +29,8 @@ public record RecommendQueryParam(
                 filter.getAgeCondition().isPermitExcessAge(),
                 filter.getDistanceCondition().getMaxIncludeDistanceFromMe(),
                 filter.getDistanceCondition().isPermitExcessDistance(),
-                point.getLatitude(),
-                point.getLongitude(),
+                point.getLatitude().doubleValue(),
+                point.getLongitude().doubleValue(),
                 size
         );
     }
